+++ lib/Doctrine/ORM/Tools/DisconnectedClassMetadataFactory.php
@@ -38,39 +38,6 @@ use Doctrine\ORM\Mapping\ClassMetadataInfo;
  */
 class DisconnectedClassMetadataFactory extends ClassMetadataFactory
 {
-    /**
-     * @override
-     */
-    protected function newClassMetadataInstance($className)
-    {
-        $metadata = new ClassMetadataInfo($className);
-        if (strpos($className, "\\") !== false) {
-            $metadata->namespace = strrev(substr( strrev($className), strpos(strrev($className), "\\")+1 ));
-        } else {
-            $metadata->namespace = "";
-        }
-        return $metadata;
-    }
-
-    /**
-     * Validate runtime metadata is correctly defined.
-     *
-     * @param ClassMetadata $class
-     * @param ClassMetadata $parent
-     */
-    protected function validateRuntimeMetadata($class, $parent)
-    {
-        // validate nothing
-    }
-
-    /**
-     * @override
-     */
-    protected function getParentClasses($name)
-    {
-        return array();
-    }
-
     public function getReflectionService()
     {
         return new \Doctrine\Common\Persistence\Mapping\StaticReflectionService;
